# backend/openapi.yaml
openapi: 3.0.3
info:
  title: ft_transcendence API
  description: >
    API REST du backend ft_transcendence (Fastify v5 + TypeScript).
    Auth JWT (access token Bearer 15 min) + refresh token en cookie httpOnly
    (SameSite=Strict). Le cookie est posé Path=/auth par le backend et réécrit en
    Path=/api/auth par le proxy Vite pour le navigateur (invariant I4). Chat temps réel
    hors-OpenAPI via WebSocket `/ws/chat` (documenté en fin de fichier pour mémoire).
  version: 0.9.0
  contact:
    name: ft_transcendence team
servers:
  # Origine applicative PRINCIPALE (navigateur) : tout passe par le proxy Vite same-origin.
  # Les chemins ci-dessous sont relatifs à cette base — /auth/login s'appelle donc
  # https://localhost:5173/api/auth/login côté navigateur.
  - url: https://localhost:5173/api
    description: App proxifiée (origine unique navigateur, I4)
  # Backend en direct : diagnostic, tests e2e Python et curl uniquement (hors navigateur app).
  - url: https://localhost:3000
    description: Backend direct (diagnostic / tests, HTTPS certificat auto-signé)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token signé (payload `{ sub }`). Les tempTokens 2FA
        (`{ sub, pending: 'totp' }`) sont refusés sur les routes protégées.
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Unauthorized
      required: [error]
    ValidationError:
      type: object
      description: Erreur de validation Zod (issues brutes).
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
                example: too_small
              path:
                type: array
                items:
                  type: string
              message:
                type: string
                example: au moins 8 caractères
      required: [errors]
    Ok:
      type: object
      properties:
        ok:
          type: boolean
          example: true
      required: [ok]

    # ---------- User ----------
...
    FriendSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pseudo:
          type: string
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true

    # ---------- Friendship ----------
    Friendship:
      type: object
      properties:
        id:
          type: string
          format: uuid
        requesterId:
          type: string
          format: uuid
        addresseeId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, accepted]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ---------- Message ----------
...
    Game:
      type: object
      description: Un jeu supporté par la plateforme.
      properties:
        id:
          type: string
          description: Slug du jeu (PK).
          example: cs2
        name:
          type: string
          example: Counter-Strike 2
        requiredProvider:
          type: string
          enum: [riot, steam, epic, chess_com]
          description: Provider d'identifiant in-game requis pour s'inscrire à un match de ce jeu.
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, requiredProvider, isActive, createdAt]
...
  responses:
...
    InternalError:
      description: Erreur interne
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /friends:
    post:
      tags: [Friends]
      summary: Envoyer une demande d'ami
      description: >
        Logique 4-cas : pas de relation → pending (201) ; demande inverse
        existante → auto-accept (200) ; même sens pending → 400 ;
        déjà amis → 400.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                addresseeId:
                  type: string
                  format: uuid
              required: [addresseeId]
      responses:
        '201':
          description: Demande créée (pending)
          content:
            application/json:
              schema:
                type: object
                properties:
                  friendship:
                    $ref: '#/components/schemas/Friendship'
        '200':
          description: Demande inverse auto-acceptée
          content:
            application/json:
              schema:
                type: object
                properties:
                  friendship:
                    $ref: '#/components/schemas/Friendship'
        '400':
          description: Soi-même, déjà demandé ou déjà amis
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Error'
                  - $ref: '#/components/schemas/ValidationError'
        '404':
          description: User cible introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Friends]
      summary: Liste de mes amis
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  friends:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/FriendSummary'
                        - type: object
                          properties:
                            since:
                              type: string
                              format: date-time
        '500':
          $ref: '#/components/responses/InternalError'
  /friends/requests:
    get:
      tags: [Friends]
      summary: Demandes d'amis reçues (pending)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  requests:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        sentAt:
                          type: string
                          format: date-time
                        from:
                          $ref: '#/components/schemas/FriendSummary'
        '500':
          $ref: '#/components/responses/InternalError'
  /friends/{id}/accept:
    post:
      tags: [Friends]
      summary: Accepter une demande
      description: Seul l'addressee peut accepter.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Demande acceptée
          content:
            application/json:
              schema:
                type: object
                properties:
                  friendship:
                    $ref: '#/components/schemas/Friendship'
        '400':
          description: >
            Param :id non-uuid (ValidationError) ou déjà amis — status non pending (Error)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationError'
                  - $ref: '#/components/schemas/Error'
        '403':
          description: Non autorisé (pas l'addressee)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Amitié introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /friends/{id}/reject:
    post:
      tags: [Friends]
      summary: Refuser une demande
      description: Seul l'addressee peut refuser. La ligne est supprimée.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Demande refusée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '400':
          description: >
            Param :id non-uuid (ValidationError) ou status non pending (Error)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationError'
                  - $ref: '#/components/schemas/Error'
        '403':
          description: Non autorisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Amitié introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /friends/{id}:
    delete:
      tags: [Friends]
      summary: Supprimer un ami
      description: >
        Unfriend d'une amitié acceptée. Les deux parties (requester ou
        addressee) peuvent supprimer.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Amitié supprimée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '400':
          description: >
            Param :id non-uuid (ValidationError) ou pas d'amitié active à supprimer (Error)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationError'
                  - $ref: '#/components/schemas/Error'
        '403':
          description: Non autorisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Amitié introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

  # =====================================================================
  # MESSAGES
  # =====================================================================
...
  /games:
    get:
      tags: [Games]
      summary: Liste les jeux supportés
      description: Renvoie tous les jeux avec `is_active = true`, triés par nom.
      responses:
        '200':
          description: Liste des jeux actifs
          content:
            application/json:
              schema:
                type: object
                properties:
                  games:
                    type: array
                    items:
                      $ref: '#/components/schemas/Game'
                required: [games]
        '500':
          $ref: '#/components/responses/InternalError'
  /games/{id}:
    get:
      tags: [Games]
      summary: Détails d'un jeu
      description: Lookup par slug (`lol`, `cs2`, `val`, `rl`, `chess`).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 50
            example: cs2
      responses:
        '200':
          description: Le jeu demandé
          content:
            application/json:
              schema:
                type: object
                properties:
                  game:
                    $ref: '#/components/schemas/Game'
                required: [game]
        '400':
          description: >
            Param :id hors bornes — 1 à 50 caractères (ValidationError). `games.id` est un
            slug `text`, pas un uuid : un slug bien formé mais inconnu reste un 404.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '404':
          description: Jeu inconnu ou inactif
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

  # =====================================================================
  # LADDERS
  # =====================================================================
...